pyproject.toml
[project]
name = "lexi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chainlit>=2.11.0",
    "chromadb>=1.5.8",
    "openai>=2.32.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sentence-transformers>=5.4.1",
]
